+++ examples/react-redux/jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    testEnvironment: "jsdom",
-    setupFilesAfterEnv: ["./src/test/setup.js"],
-};