+++ examples/react-redux/src/components/footer.jsx
@@ -1,45 +1,36 @@
-import { Component } from "react";
 import classnames from "classnames";
-import PropTypes from "prop-types";
+import { useDispatch } from "react-redux";
+import { clearCompleted } from "../store";
 
-import { SHOW_ALL, SHOW_COMPLETED, SHOW_ACTIVE } from "../constants/todo-filters";
+export default function Footer({ completedCount, activeCount, filter, hidden = false }) {
+    const dispatch = useDispatch();
 
-export default class Footer extends Component {
-    static propTypes = {
-        completedCount: PropTypes.number.isRequired,
-        activeCount: PropTypes.number.isRequired,
-        filter: PropTypes.string.isRequired,
-        onClearCompleted: PropTypes.func.isRequired,
-    };
-
-    render() {
-        const { completedCount, onClearCompleted, activeCount, filter } = this.props;
-        return (
-            <footer className="footer" data-testid="footer">
-                <span className="todo-count">{`${activeCount} ${activeCount === 1 ? "item" : "items"} left!`}</span>
-                <ul className="filters" data-testid="footer-navigation">
-                    <li>
-                        <a className={classnames({ selected: filter === SHOW_ALL })} href={`#${SHOW_ALL}`}>
-                            All
-                        </a>
-                    </li>
-                    <li>
-                        <a className={classnames({ selected: filter === SHOW_ACTIVE })} href={`#${SHOW_ACTIVE}`}>
-                            Active
-                        </a>
-                    </li>
-                    <li>
-                        <a className={classnames({ selected: filter === SHOW_COMPLETED })} href={`#${SHOW_COMPLETED}`}>
-                            Completed
-                        </a>
-                    </li>
-                </ul>
-                {completedCount > 0 ? (
-                    <button className="clear-completed" onClick={onClearCompleted}>
-                        Clear completed
-                    </button>
-                ) : null}
-            </footer>
-        );
-    }
+    return (
+        <footer className="footer" data-testid="footer" hidden={hidden}>
+            <span className="todo-count">{`${activeCount} ${activeCount === 1 ? "item" : "items"} left!`}</span>
+            <ul className="filters" data-testid="footer-navigation">
+                <li>
+                    <a
+                        className={classnames({ selected: filter === "/" || filter === "/all" })}
+                        href="#/"
+                    >
+                        All
+                    </a>
+                </li>
+                <li>
+                    <a className={classnames({ selected: filter === "/active" })} href="#/active">Active</a>
+                </li>
+                <li>
+                    <a className={classnames({ selected: filter === "/completed" })} href="#/completed">Completed</a>
+                </li>
+            </ul>
+            <button
+                className="clear-completed"
+                hidden={completedCount === 0}
+                onClick={() => dispatch(clearCompleted())}
+            >
+                Clear completed
+            </button>
+        </footer>
+    );
 }