+++ examples/react-redux/src/components/item.jsx
@@ -1,62 +1,54 @@
-import { PureComponent } from "react";
+import { memo, useState } from "react";
 import classnames from "classnames";
-import PropTypes from "prop-types";
+import { useDispatch } from "react-redux";
 import TextInput from "./text-input";
+import { editTodo, deleteTodo, toggleTodo } from "../store";
 
-export default class Item extends PureComponent {
-    static propTypes = {
-        todo: PropTypes.object.isRequired,
-        editTodo: PropTypes.func.isRequired,
-        deleteTodo: PropTypes.func.isRequired,
-        toggleTodo: PropTypes.func.isRequired,
-        index: PropTypes.number.isRequired,
+function Item({ todo }) {
+    const dispatch = useDispatch();
+    const [editing, setEditing] = useState(false);
+
+    const handleSave = (text) => {
+        if (text.length === 0) dispatch(deleteTodo(todo.id));
+        else dispatch(editTodo({ id: todo.id, text }));
+        setEditing(false);
     };
 
-    state = {
-        editing: false,
-    };
+    const handleCancel = () => setEditing(false);
 
-    handleDoubleClick = () => {
-        this.setState({ editing: true });
-    };
-
-    handleSave = (id, text) => {
-        if (text.length === 0)
-            this.props.deleteTodo(id);
-        else
-            this.props.editTodo(id, text);
-
-        this.setState({ editing: false });
-    };
-
-    render() {
-        const { todo, toggleTodo, deleteTodo, index } = this.props;
-
-        let element;
-        if (this.state.editing) {
-            element = <TextInput text={todo.text} editing={this.state.editing} onSave={(text) => this.handleSave(todo.id, text)} />;
-        } else {
-            element = (
-                <div className="view">
-                    <input className="toggle" type="checkbox" data-testid="todo-item-toggle" checked={todo.completed} onChange={() => toggleTodo(todo.id)} />
-                    <label onDoubleClick={this.handleDoubleClick} data-testid="todo-item-label">
-                        {todo.text}
-                    </label>
-                    <button className="destroy" data-testid="todo-item-button" onClick={() => deleteTodo(todo.id)} />
-                </div>
-            );
-        }
-
-        return (
-            <li
-                className={classnames({
-                    completed: todo.completed,
-                    editing: this.state.editing,
-                })}
-                data-testid="todo-item"
-            >
-                {element}
-            </li>
-        );
-    }
+    return (
+        <li
+            className={classnames({ completed: todo.completed, editing })}
+            data-testid="todo-item"
+        >
+            <div className="view">
+                <input
+                    className="toggle"
+                    type="checkbox"
+                    data-testid="todo-item-toggle"
+                    checked={todo.completed}
+                    onChange={() => dispatch(toggleTodo(todo.id))}
+                />
+                <label onDoubleClick={() => setEditing(true)} data-testid="todo-item-label">
+                    {todo.text}
+                </label>
+                <button
+                    className="destroy"
+                    data-testid="todo-item-button"
+                    aria-label="Delete todo"
+                    onClick={() => dispatch(deleteTodo(todo.id))}
+                />
+            </div>
+            {editing && (
+                <TextInput
+                    text={todo.text}
+                    editing
+                    onSave={handleSave}
+                    onCancel={handleCancel}
+                />
+            )}
+        </li>
+    );
 }
+
+export default memo(Item);