+++ examples/react-redux/src/components/main.jsx
@@ -1,43 +1,52 @@
-import { Component } from "react";
-import classnames from "classnames";
-import PropTypes from "prop-types";
+import { useDispatch, useSelector } from "react-redux";
+import { useLocation } from "react-router-dom";
+
 import Item from "./item";
 import Footer from "./footer";
-export default class Main extends Component {
-    static propTypes = {
-        todos: PropTypes.array.isRequired,
-        location: PropTypes.object.isRequired,
-        visibleTodos: PropTypes.array.isRequired,
-        completedCount: PropTypes.number.isRequired,
-        activeCount: PropTypes.number.isRequired,
-        editTodo: PropTypes.func.isRequired,
-        deleteTodo: PropTypes.func.isRequired,
-        toggleTodo: PropTypes.func.isRequired,
-        toggleAll: PropTypes.func.isRequired,
-        clearCompleted: PropTypes.func.isRequired,
-    };
+import {
+    selectCompletedCount,
+    selectTodos,
+    selectVisibleTodos,
+    toggleAll,
+} from "../store";
 
-    render() {
-        const { todos, editTodo, deleteTodo, toggleTodo, toggleAll, clearCompleted, location, visibleTodos, completedCount, activeCount } = this.props;
+export default function Main() {
+    const dispatch = useDispatch();
+    const location = useLocation();
+    const todos = useSelector(selectTodos);
+    const visibleTodos = useSelector(selectVisibleTodos(location.pathname));
+    const completedCount = useSelector(selectCompletedCount);
+    const activeCount = todos.length - completedCount;
+    const empty = todos.length === 0;
 
-        if (todos.length === 0)
-            return null;
-
-        return (
-            <main className="main" data-testid="main">
+    return (
+        <>
+            <main className="main" data-testid="main" hidden={empty}>
                 <div className="toggle-all-container">
-                    <input className="toggle-all" type="checkbox" data-testid="toggle-all" checked={completedCount === todos.length} onChange={toggleAll} />
+                    <input
+                        id="toggle-all"
+                        className="toggle-all"
+                        type="checkbox"
+                        data-testid="toggle-all"
+                        checked={!empty && completedCount === todos.length}
+                        onChange={() => dispatch(toggleAll())}
+                    />
                     <label className="toggle-all-label" htmlFor="toggle-all">
                         Toggle All Input
                     </label>
                 </div>
-                <ul className={classnames("todo-list")} data-testid="todo-list">
-                    {visibleTodos.map((todo, index) => (
-                        <Item key={todo.id} todo={todo} editTodo={editTodo} deleteTodo={deleteTodo} toggleTodo={toggleTodo} index={index} />
+                <ul className="todo-list" data-testid="todo-list">
+                    {visibleTodos.map((todo) => (
+                        <Item key={todo.id} todo={todo} />
                     ))}
                 </ul>
-                <Footer completedCount={completedCount} activeCount={activeCount} filter={location.pathname} onClearCompleted={clearCompleted} />
             </main>
-        );
-    }
+            <Footer
+                completedCount={completedCount}
+                activeCount={activeCount}
+                filter={location.pathname}
+                hidden={empty}
+            />
+        </>
+    );
 }