+++ examples/react/src/index.js
@@ -1,16 +1,14 @@
-import React from "react";
-import { render } from "react-dom";
+import { createRoot } from "react-dom/client";
 import { HashRouter, Route, Routes } from "react-router-dom";
 
 import { App } from "./todo/app";
 import "todomvc-app-css/index.css";
 import "todomvc-common/base.css";
 
-render(
+createRoot(document.getElementById("root")).render(
     <HashRouter>
         <Routes>
             <Route path="*" element={<App />} />
         </Routes>
-    </HashRouter>,
-    document.getElementById("root")
+    </HashRouter>
 );