+++ examples/svelte/src/Footer.svelte
@@ -1,18 +1,8 @@
 <script>
-    import { createEventDispatcher } from 'svelte';
-
-    export let numActive;
-    export let currentFilter;
-    export let numCompleted;
-
-    const dispatch = createEventDispatcher();
-
-    function removeCompletedItems(e) {
-        dispatch('removeCompletedItems');
-    }
+    let { numActive, currentFilter, numCompleted, onRemoveCompletedItems, hidden = false } = $props();
 </script>
 
-<footer class="footer">
+<footer class="footer" {hidden}>
     <span class="todo-count">
         <strong>{numActive}</strong>
         {numActive === 1 ? "item" : "items"} left
@@ -24,7 +14,7 @@
         <li><a class:selected={currentFilter === "completed"} href="#/completed">Completed</a></li>
     </ul>
 
-    {#if numCompleted}
-        <button class="clear-completed" on:click={removeCompletedItems}> Clear completed </button>
-    {/if}
+    <button class="clear-completed" onclick={onRemoveCompletedItems} hidden={numCompleted === 0}>
+        Clear completed
+    </button>
 </footer>